Extract IPAM resolver and cover compute allocation
  Move network/IPAM resolution into ipam.go (ipamResolver) and add an
  end-to-end AllocateComputeResource test plus IPAM unit tests. Share
  test helpers via k8stest, assert typed-error propagation in the ViVnfm
  adapters, and run the suites in parallel.
